Integrate the new RegressionData struct
The new struct was integrated into the FormulaRegressionBuilder.
It turns out that we will not be able to make FormulaRegressionBuilder
implement the Copy trait as the formula field also does not.
However cloning a single String should have less of a performance impact
compared to cloning the entire data HashMap.
Furthermore the type of the formula field was changed to Cow<'a, str>.