Fully replace cephes special functions with new implementation
We can compute the p-values using the cdf for a Student's T
distribution with location 0 and scale 1.0.

This lets us replace the `stdtr` function that was originally
ported from the cephes library with an implementation
adapted from the statrs library, that is well documented
and doesn't trigger clippy's many_single_char_names lint.